Add check for validator_registry and validator_balances length
