+++ tests/beacon/state_machines/test_validation.py
@@ -20,7 +20,7 @@ from eth.beacon.types.states import BeaconState
 
 from eth._utils import bls
 
-from tests.beacon.helpers import mock_active_validator_record
+from tests.beacon.helpers import mock_validator_record
 from tests.beacon.test_helpers import (
     get_sample_shard_committees_at_slots,
 )
@@ -46,13 +46,18 @@ def test_validate_proposer_signature(
         sample_beacon_state_params,
         sample_shard_committee_params,
         beacon_chain_shard_number,
-        epoch_length):
+        epoch_length,
+        max_deposit):
 
     state = BeaconState(**sample_beacon_state_params).copy(
-        validator_registry=[
-            mock_active_validator_record(proposer_pubkey)
+        validator_registry=tuple(
+            mock_validator_record(proposer_pubkey)
             for _ in range(10)
-        ],
+        ),
+        validator_balances=tuple(
+            max_deposit
+            for _ in range(10)
+        ),
         shard_committees_at_slots=get_sample_shard_committees_at_slots(
             num_slot=128,
             num_shard_committee_per_slot=10,
